requirements.txt
# Web Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
jinja2==3.1.3
itsdangerous==2.1.2

# Database
sqlalchemy==2.0.25
aiosqlite==0.19.0
alembic==1.18.4

# File Parsing
pymupdf==1.23.8
python-docx==1.1.0
olefile==0.46
Pillow>=10.0.0

# Google Gemini
google-genai>=1.0.0

# Validation & Settings
pydantic>=2.6.0
pydantic-settings==2.1.0
email-validator==2.1.0

# Async Support
aiofiles==23.2.1

# Rate Limiting
slowapi==0.1.9

# Utilities
python-dateutil==2.8.2
python-dotenv==1.0.0
httpx==0.26.0
bcrypt==4.1.2

# OAuth
authlib==1.3.0

# Email
aiosmtplib==3.0.1
